remade GameItems according to new Phantom concept
