+++ src/main/java/net/sknv/engine/SkyBox.java
@@ -1,6 +1,6 @@
 package net.sknv.engine;
 
-import net.sknv.engine.entities.GameItemMesh;
+import net.sknv.engine.entities.AbstractGameItem;
 import net.sknv.engine.graph.*;
 import org.joml.Matrix4f;
 import org.joml.Vector3f;
@@ -10,7 +10,7 @@ import static org.lwjgl.opengl.GL13.GL_TEXTURE0;
 import static org.lwjgl.opengl.GL13.glActiveTexture;
 import static org.lwjgl.opengl.GL30.glBindVertexArray;
 
-public class SkyBox extends GameItemMesh {
+public class SkyBox extends AbstractGameItem {
 
     private Matrix4f projectionMatrix;
     private Vector3f ambientLight;
