+++ src/main/java/net/sknv/engine/entities/Terrain.java
@@ -13,7 +13,7 @@ import static org.lwjgl.stb.STBImage.*;
 
 public class Terrain {
 
-    private final GameItemMesh[] gameItems;
+    private final Phantom[] gameItems;
 
     private final int terrainSize;
 
@@ -43,7 +43,7 @@ public class Terrain {
      */
     public Terrain(int terrainSize, float scale, float minY, float maxY, String heightMapFile, String textureFile, int textInc) throws Exception {
         this.terrainSize = terrainSize;
-        gameItems = new GameItemMesh[terrainSize * terrainSize];
+        gameItems = new Phantom[terrainSize * terrainSize];
 
         ByteBuffer buf = null;
         int width;
@@ -73,7 +73,7 @@ public class Terrain {
                 float xDisplacement = (col - ((float) terrainSize - 1) / (float) 2) * scale * HeightMapMesh.getXLength();
                 float zDisplacement = (row - ((float) terrainSize - 1) / (float) 2) * scale * HeightMapMesh.getZLength();
 
-                GameItemMesh terrainBlock = new GameItemMesh(heightMapMesh.getMesh());
+                Phantom terrainBlock = new Phantom(heightMapMesh.getMesh());
                 terrainBlock.setScale(scale);
                 terrainBlock.setPosition(xDisplacement, 0, zDisplacement);
                 gameItems[row * terrainSize + col] = terrainBlock;
@@ -91,7 +91,7 @@ public class Terrain {
         // and check if the position is contained in that bounding box
         Box2D boundingBox = null;
         boolean found = false;
-        GameItemMesh terrainBlock = null;
+        Phantom terrainBlock = null;
         for (int row = 0; row < terrainSize && !found; row++) {
             for (int col = 0; col < terrainSize && !found; col++) {
                 terrainBlock = gameItems[row * terrainSize + col];
@@ -110,7 +110,7 @@ public class Terrain {
         return result;
     }
 
-    protected Vector3f[] getTriangle(Vector3f position, Box2D boundingBox, GameItemMesh terrainBlock) {
+    protected Vector3f[] getTriangle(Vector3f position, Box2D boundingBox, Phantom terrainBlock) {
         // Get the column and row of the heightmap associated to the current position
         float cellWidth = boundingBox.width / (float) verticesPerCol;
         float cellHeight = boundingBox.height / (float) verticesPerRow;
@@ -146,7 +146,7 @@ public class Terrain {
         return z;
     }
 
-    protected float getWorldHeight(int row, int col, GameItemMesh gameItem) {
+    protected float getWorldHeight(int row, int col, Phantom gameItem) {
         float y = heightMapMesh.getHeight(row, col);
         return y * gameItem.getScale() + gameItem.getPosition().y;
     }
@@ -168,7 +168,7 @@ public class Terrain {
      * @param terrainBlock A GameItem instance that defines the terrain block
      * @return The boundingg box of the terrain block
      */
-    private Box2D getBoundingBox(GameItemMesh terrainBlock) {
+    private Box2D getBoundingBox(Phantom terrainBlock) {
         float scale = terrainBlock.getScale();
         Vector3f position = terrainBlock.getPosition();
